Make background 16:9 aspect ratio and add zoom funtionality with bounds
The background tiles, sea, have been changed to that the propotion is roughly 16:9 so the camera can see the whole board but not empty space. Zoom functionality has been added, which includes bounds so that the empty space can't be seen and the camera is confined to the arena and closer to the centre the bigger it gets.