Add Bristol Lab robot and scripts to hold the states and behaviour of the algorithm
Added a simple prefab to represent the robot with a square to indicate that transform.right is the forward direction, which aligns with the axis that in Unity. Scripts to represent the different states of the robot have also been added to better represent the algorithm and prevent new implementations from affecting other states.