Implement detected of object in field of view
Implemented the field of view by detecting objects within range, using a collider, and then checking whether they are with the angle of the field of view. The editor script will also draw a line to the objects it can see. NOTE: There are some false negatives and positives for the border objects which might indicate another method is needed.